refactored App to Layout and added router
