+++ mqlfiddle-fe/src/components/Layout/style.css
@@ -1,0 +1,3 @@
+.content-container {
+    height: calc(100% - 64px);
+}
